add generateUserId()

+++ src/app/components/users.component.ts
@@ -12,10 +12,7 @@ export class UsersComponent implements OnInit {
   users: any;
   errorMessage: any;
 
-
   constructor(private dataService: DataService) { }
-
-
 
   getUsers() {
     const usersUrl: string = 'http://localhost:3000/users';
@@ -29,9 +26,7 @@ export class UsersComponent implements OnInit {
   addUser() {
     const addUserUrl: string = 'http://localhost:3000/users';
 
-    const id = this.users[this.users.length - 1].id + 1;
-
-    const data: User = { id: id, name: 'user', email: 'user@internet.com', telephone_number: '07955369541' }
+    const data: User = { id: this.generateUserId(), name: 'user', email: 'user@internet.com', telephone_number: '07955369541' }
 
     this.dataService.post(addUserUrl, data)
       .then(user => this.users.push(user))
@@ -39,12 +34,22 @@ export class UsersComponent implements OnInit {
   }
 
   deleteUser() {
+    if (this.users.length == 0) return;
 
-    let last = this.users[this.users.length - 1].id
+    let last = this.users[this.users.length - 1].id;
 
     this.dataService.delete(last)
       .then(() => this.users = this.users.filter((user: any) => user.id !== last))
       .catch(error => this.errorMessage = `${error}: Could not delete user. Please try again.`)
+  }
+
+  private generateUserId(): number {
+    let id: number = 1
+
+    if (this.users.length !== 0) {
+      id = this.users[this.users.length - 1].id + 1;
+    }
+    return id;
   }
 
   ngOnInit() {
